feat(municipality-restructuring): applied PR feedback and cleaned up a bit;
